RH-1354 #comment Fix bug about add patient and Refactor add treatment feature

+++ src/groovy/com/ratchethealth/client/CareGiver.groovy
@@ -1,8 +1,8 @@
 package com.ratchethealth.client
 
 class CareGiver {
-    int medicalRecordId
-    int careGiverId
+    Long medicalRecordId
+    Long careGiverId
     String email
     String firstName
     String lastName
