+++ src/groovy/com/ratchethealth/client/Treatment.groovy
@@ -1,12 +1,12 @@
 package com.ratchethealth.client
 
 class Treatment {
-    int clientId
-    int patientId
-    int treatmentId
-    int staffIds
+    Long clientId
+    Long patientId
+    Long treatmentId
+    Long staffIds
     Long surgeryTime
-    int groupId
+    Long groupId
     String ecFirstName
     String ecLastName
     String relationship
